+++ app/src/main/java/opensource/cached_dupe_scanner/cache/CacheDatabaseFactory.kt
@@ -27,7 +27,8 @@ val CACHE_DATABASE_MIGRATIONS: Array<Migration> = arrayOf(
     CacheMigrations.MIGRATION_18_19,
     CacheMigrations.MIGRATION_19_20,
     CacheMigrations.MIGRATION_20_21,
-    CacheMigrations.MIGRATION_21_22
+    CacheMigrations.MIGRATION_21_22,
+    CacheMigrations.MIGRATION_22_23
 )
 
 fun buildCacheDatabase(context: Context): CacheDatabase {